requirements.txt
pytest~=8.3.3
selene==2.0.0rc7
selenium~=4.26.1
allure-pytest
allure-python-commons~=2.13.5
python-dotenv
webdriver_manager
Appium-Python-Client
requests